+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* BPI driver to use new login page with iFrame structure
 
 ## [0.1.7] - 2024-12-21